DOSGI-156 Fix for failures when run with OSGi CTs
This commit provides fixes for most, possibly all of the failures experienced with the OSGi Remote Services and Remote Service Admin CTs. 
There is a small number of CT failures that relate to the arrival of events, but it looks like there might be an error in the CT logic around this. I have raised an issue against the CT for this. 
I added unit tests for most of the behaviour that I changed, but am planning to add more unit tests in subsequent commits.



git-svn-id: https://svn.apache.org/repos/asf/cxf/dosgi/trunk@1442924 13f79535-47bb-0310-9956-ffa450edef68